feat(web): Externalizing http server
