+++ module/web-aws-lambda/__index__.ts
@@ -1,2 +1,2 @@
-export * from './src/application.ts';
+export * from './src/handler.ts';
 export * from './src/util.ts';
